+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.service.ReservationService;
-import roomescape.service.dto.reservation.request.ReservationRequest;
-import roomescape.service.dto.reservation.response.ReservationResponse;
+import roomescape.controller.reservation.request.ReservationRequest;
+import roomescape.service.reservation.ReservationService;
+import roomescape.service.reservation.response.ReservationResponse;
 
 @RequestMapping("/reservations")
 @RestController
@@ -24,15 +24,15 @@ public final class ReservationController {
     }
 
     @PostMapping
-    public ResponseEntity<ReservationResponse> save(@RequestBody ReservationRequest reservationRequest) {
-        final ReservationResponse reservationResponse = reservationService.save(reservationRequest);
-        return ResponseEntity.ok(reservationResponse);
+    public ResponseEntity<ReservationResponse> save(@RequestBody ReservationRequest request) {
+        final ReservationResponse response = reservationService.save(request.toServiceRequest());
+        return ResponseEntity.ok(response);
     }
 
     @GetMapping
     public ResponseEntity<List<ReservationResponse>> readAll() {
-        final List<ReservationResponse> reservationResponses = reservationService.readAll();
-        return ResponseEntity.ok(reservationResponses);
+        final List<ReservationResponse> responses = reservationService.readAll();
+        return ResponseEntity.ok(responses);
     }
 
     @DeleteMapping("/{id}")
