+++ src/main/java/roomescape/service/reservation/request/ReservationServiceRequest.java
@@ -1,10 +1,10 @@
-package roomescape.service.dto.reservation.request;
+package roomescape.service.reservation.request;
 
 import java.time.LocalDate;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
 
-public record ReservationRequest(
+public record ReservationServiceRequest(
         String name,
         LocalDate date,
         Long timeId) {
