+++ src/main/java/roomescape/service/reservationtime/request/ReservationTimeServiceRequest.java
@@ -1,9 +1,9 @@
-package roomescape.service.dto.reservationtime.request;
+package roomescape.service.reservationtime.request;
 
 import java.time.LocalTime;
 import roomescape.model.ReservationTime;
 
-public record ReservationTimeRequest(LocalTime startAt) {
+public record ReservationTimeServiceRequest(LocalTime startAt) {
 
     public ReservationTime toTime() {
         return ReservationTime.ofWithoutId(this.startAt);
